added simplejson to requirements; added a utility print_dict using json to generic.py; split out read Bruker tests from test_utilities.py
