+++ .gitignore
@@ -1,3 +1,4 @@
+.settings/
 .project
 .pydevproject
 *.mat
