+++ nPYc/utilities/generic.py
@@ -1,6 +1,8 @@
 """
 Generic Utility functions
 """
+import simplejson as json
+
 def removeDuplicateColumns(df):
 	"""
 	Removes duplicate columns from the passed dataframe
@@ -29,3 +31,16 @@ def removeTrailingColumnNumbering(column_list):
 			tmp.append(s)
 
 	return tmp
+
+def print_dict(message, dict_obj):  
+    """
+    Use JSON to print out a dict in an easily readable format
+    """
+    my_complex_dict = json.dumps(dict_obj, indent=4, default=str)
+    print(f"{message}:\n\t{my_complex_dict}")  
+ 
+def print_dict_list(message, list_obj):
+	json_list = json.dumps(list_obj, indent=4, default=str)
+	print(f"{message}:\n\t{json_list}")
+	
+
