+++ requirements.txt
@@ -15,3 +15,4 @@ seaborn>=0.8.1
 setuptools>=39.1.0
 statsmodels>=0.9.0
 pyChemometrics>=0.1
+simplejson
